[feature] Safe Args in Navigation Compose
