+++ shared/src/commonMain/kotlin/com.santimattius.kmp.entertainment/navigation/AppNavigation.kt
@@ -14,7 +14,6 @@ import com.santimattius.kmp.entertainment.feature.movie.home.MoviesRoute
 import com.santimattius.kmp.entertainment.feature.splash.SplashScreen
 import com.santimattius.kmp.entertainment.feature.tvshow.detail.TvShowDetailRoute
 import com.santimattius.kmp.entertainment.feature.tvshow.home.TvShowRoute
-import kotlinx.serialization.Serializable
 
 
 @Composable
@@ -68,23 +67,3 @@ fun AppNavigation(
         }
     }
 }
-
-// TODO: move this definitions
-
-@Serializable
-data object Splash
-
-@Serializable
-data object Movie
-
-@Serializable
-data class MovieDetail(val id: Long)
-
-@Serializable
-data object TvShow
-
-@Serializable
-data class TvShowDetail(val id: Long)
-
-@Serializable
-data object Favorite
